refactor: :recycle: 更新数据类型

+++ internal/api/common/resp.go
@@ -40,7 +40,7 @@ type ResponseValidationStruct struct {
 type ResponsePaginationStruct struct {
 	Page     int         `json:"page" example:"1"`       // 当前页码
 	PageSize int         `json:"page_size" example:"10"` // 每页大小
-	Total    int64       `json:"total" example:"100"`    // 总记录数
+	Total    uint16      `json:"total" example:"100"`    // 总记录数
 	Data     interface{} `json:"data"`                   // 数据列表
 }
