+++ internal/database/interfaces/notify.go
@@ -12,25 +12,25 @@ type NotifyRepository interface {
 	Create(ctx context.Context, channel *notify.Data) error
 
 	// GetByID 根据ID获取通知渠道
-	GetByID(ctx context.Context, id int64) (*notify.Data, error)
+	GetByID(ctx context.Context, id uint16) (*notify.Data, error)
 
 	// Update 更新通知渠道
 	Update(ctx context.Context, channel *notify.Data) error
 
 	// Delete 删除通知渠道
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint16) error
 
 	// List 获取通知渠道列表
-	List(ctx context.Context, offset, limit int) (*[]notify.Data, error)
+	List(ctx context.Context, offset, limit uint16) (*[]notify.Data, error)
 
 	// Count 获取通知渠道总数
-	Count(ctx context.Context) (int64, error)
+	Count(ctx context.Context) (uint16, error)
 
 	// 根据任务ID获取通知渠道列表
-	GetByTaskID(ctx context.Context, taskID int64) (*[]notify.Data, error)
+	GetByTaskID(ctx context.Context, taskID uint16) (*[]notify.Data, error)
 
 	// 添加通知渠道与任务的关联
-	AddTaskRelation(ctx context.Context, notifyID, taskID int64) error
+	AddTaskRelation(ctx context.Context, notifyID, taskID uint16) error
 }
 
 // NotificationTemplateRepository 通知模板数据访问接口
@@ -39,23 +39,23 @@ type NotifyTemplateRepository interface {
 	Create(ctx context.Context, template *notify.Template) error
 
 	// GetByID 根据ID获取通知模板
-	GetByID(ctx context.Context, id int64) (*notify.Template, error)
+	GetByID(ctx context.Context, id uint16) (*notify.Template, error)
 
 	// Update 更新通知模板
 	Update(ctx context.Context, template *notify.Template) error
 
 	// Delete 删除通知模板
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint16) error
 
 	// List 获取通知模板列表
-	List(ctx context.Context, offset, limit int) (*[]notify.Template, error)
+	List(ctx context.Context, offset, limit uint16) (*[]notify.Template, error)
 
 	// Count 获取通知模板总数
-	Count(ctx context.Context) (int64, error)
+	Count(ctx context.Context) (uint16, error)
 
 	// 根据任务ID获取通知模板
-	GetByTaskID(ctx context.Context, taskID int64) (*notify.Template, error)
+	GetByTaskID(ctx context.Context, taskID uint16) (*notify.Template, error)
 
 	// 添加通知模板与任务的关联
-	AddTaskRelation(ctx context.Context, templateID, taskID int64) error
+	AddTaskRelation(ctx context.Context, templateID, taskID uint16) error
 }
