+++ internal/models/sub/fetch.go
@@ -3,7 +3,7 @@ package sub
 type FetchResult struct {
 	Type      string `json:"sub_type"`
 	NodeCount int    `json:"node_count"`
-	Size      int64  `json:"size"`
+	Size      uint32 `json:"size"`
 	Duration  string `json:"duration"`
 	Error     string `json:"error"`
 }
