+++ internal/modules/parser/singbox/singbox.go
@@ -1,6 +1,6 @@
 package singbox
 
-func Parse(content *[]byte, sublinkID int64) (int, error) {
+func Parse(content *[]byte, sublinkID uint16) (int, error) {
 
 	return 0, nil
 }
